Organize Amazon release deprecations (#399)
